Crush yet-another wx2.9 wxString::Format assert. Fixes bug #3429429
git-svn-id: https://codelite.svn.sourceforge.net/svnroot/codelite/trunk@5177 9da81c78-c036-0410-9e1f-a2b0375e4b5a

+++ UnitTestCPP/unittestcppoutputparser.cpp
@@ -149,5 +149,5 @@ void TestSummary::PrintSelf()
 {
 	wxPrintf(wxT("Total tests            : %d\n"), totalTests);
 	wxPrintf(wxT("Total errors           : %d\n"), errorCount);
-	wxPrintf(wxT("Total error lines found: %d\n"), errorLines.size());
+	wxPrintf(wxT("Total error lines found: %u\n"), (unsigned int)errorLines.size());
 }
